+++ src/lib/components/buttons/flat_white.tsx
@@ -77,6 +77,7 @@ const styles = StyleSheet.create({
     alignItems: "center",
     borderColor: colors["gray-regular"],
     borderWidth: 1,
+    flex: 1,
   },
   text: {
     fontFamily: "AGaramondPro-Regular",