+++ src/lib/components/consignments/setup/welcome.tsx
@@ -1,6 +1,12 @@
 import * as React from "react"
 import * as Relay from "react-relay"
-import Button from "../../buttons/inverted_button"
+
+import Button from "../../buttons/flat_white"
+import Circle from "../components/circle_image"
+import {
+  BodyText as P,
+  LargeHeadline,
+} from "../typography"
 import Info from "./info"
 
 import {
@@ -8,7 +14,7 @@ import {
   NavigatorIOS,
   Route,
   ScrollView,
-  Text,
+  StyleSheet,
   TouchableHighlight,
   View,
   ViewProperties,
@@ -16,7 +22,7 @@ import {
 
 interface Props extends ViewProperties {
   navigator: NavigatorIOS,
-  route: Route // this gets set by NavigatorIOS
+  route: Route
 }
 
 export default class Welcome extends React.Component<Props, any> {
@@ -28,12 +34,22 @@ export default class Welcome extends React.Component<Props, any> {
 
   render() {
     return  (
-      <View style={{marginTop: 20, backgroundColor: "green"}}>
-        <Text>Hello this is the welcome</Text>
-        <View style={{height: 26, width: 320, marginTop: 20}}>
-          <Button text="Go" onPress={this.goTapped}/>
+      <ScrollView style={{backgroundColor: "black"}}>
+        <View style={{flex: 1, paddingTop: 40, alignItems: "center"}}>
+          <LargeHeadline>Sell works from your collection through our partner network</LargeHeadline>
+
+          <View style={{width: 300, alignItems: "center", marginTop: 20}}>
+            <Circle source={require("../images/email.png")} />
+            <P>Sell work from your collection through our partner network.</P>
+            <Circle source={require("../images/hammer.png")} />
+            <P>Get your work placed in an upcoming sale.</P>
+          </View>
+
+          <View style={{height: 43, width: 320, marginTop: 20}}>
+            <Button text="GET STARTED" onPress={this.goTapped}/>
+          </View>
         </View>
-      </View>
+      </ScrollView>
     )
   }
 }